+++ Examples/IntegerSet/IntegerSet_04_Vector_MoveSemantics.cpp
@@ -25,7 +25,7 @@ namespace IntegerSetVector_MoveSemantics {
 
     // getter
     int IntegerSet::size() const {
-        return m_set.size();
+        return static_cast<int> (m_set.size());
     }
 
     bool IntegerSet::isEmpty() const {
